requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-Session==0.4.0
python-dotenv==1.0.0
pymongo==4.3.3
SQLAlchemy==2.0.5
Werkzeug==2.2.3
WTForms==3.0.1
googlemaps==4.10.0
polyline==2.0.0
pyopenssl==23.0.0
fpdf2==2.7.4
geopy==2.3.0
requests==2.28.2
Pillow==9.4.0
email-validator==1.3.1
gunicorn==20.1.0